Make on Linux the debug version compile and link
However, the game starts, but I get some debug asserts I don't get on Windows and it crashes right after that. It does not create the logs and probably no leaks report, either.

Make configure defining the right preprocessor directives for the debug built.
The debug built on Windows has defined _DEBUG, _PLAYTEST, _DEBUGTOOLS, and
_DEBUG_MEMORY, but it does not define
_BFR_, that is only for the final build.

Make file fix, the debug and logging version needs Timer.cpp built.
ctp2_code\Makefile.am
ctp2_code\ctp\Makefile.am

On a debug assert the game should break and not abort
ctp2_code\ctp\ctp2_utils\c3debug.cpp

On a break point the game should break
ctp2_code\ctp\debugtools\breakpoint.h

Merge code from the Linux branch in and make it compile and link
ctp2_code\ctp\civapp.cpp
ctp2_code\ctp\ctp2_utils\c3debug.cpp
ctp2_code\ctp\ctp2_utils\c3debug.h
ctp2_code\ctp\debugtools\debug.c
ctp2_code\ctp\debugtools\debug.h
ctp2_code\ctp\debugtools\debugassert.cpp
ctp2_code\ctp\debugtools\debugassert.h
ctp2_code\ctp\debugtools\debugcallstack.cpp
ctp2_code\ctp\debugtools\debugcallstack.h
ctp2_code\ctp\debugtools\debugexception.cpp
ctp2_code\ctp\debugtools\debugexception.h
ctp2_code\ctp\debugtools\debugmemory.cpp
ctp2_code\ctp\debugtools\debugmemory.h
ctp2_code\ctp\debugtools\log.cpp
ctp2_code\ctp\debugtools\log.h
ctp2_code\gs\slic\slic.y
ctp2_code\os\nowin32\windows.h
ctp2_code\ui\aui_common\auidebug.h
ctp2_code\ui\aui_ctp2\keypress.cpp

Cleanup and white space
ctp2_code\ctp\ctp2_utils\c3mem.cpp
ctp2_code\ctp\debugtools\log_off.h
ctp2_code\ctp\debugtools\log_on.h

+++ ctp2_code/ctp/debugtools/debugcallstack.h
@@ -44,9 +44,9 @@ extern "C" {
 void DebugCallStack_Open (void);
 void DebugCallStack_Close (void);
 void DebugCallStack_Dump (LogClass log_class);
-void DebugCallStack_DumpFrom (LogClass log_class, unsigned base_pointer);
-void DebugCallStack_DumpAddress (LogClass log_class, unsigned address);
-void DebugCallStack_Save  (unsigned *call_stack, int number, unsigned long Ebp);
+void DebugCallStack_DumpFrom (LogClass log_class, size_t base_pointer);
+void DebugCallStack_DumpAddress (LogClass log_class, size_t address);
+void DebugCallStack_Save  (unsigned *call_stack, int number, size_t Ebp);
 
 void DebugCallStack_Show  (LogClass log_class, unsigned *call_stack, int number);
 void DebugCallStack_ShowToFile  (LogClass log_class, unsigned *call_stack, int number, FILE *file);
